requirements.txt
streamlit==1.28.0
web3==6.10.0
py-solc-x==1.1.1
eth-account==0.9.0
matplotlib==3.7.2
pandas==2.0.3
